transferUart/requirements.txt
pyserial==3.5
pycryptodome==3.19.0 